Move code for markers into an interface
