+++ jvm/markerapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/markerapp/app/viewmodel/VerseMarkerViewModel.kt
@@ -33,7 +33,6 @@ import org.slf4j.LoggerFactory
 import org.wycliffeassociates.otter.common.audio.AudioFile
 import org.wycliffeassociates.otter.common.device.IAudioPlayer
 import org.wycliffeassociates.otter.jvm.controls.controllers.AudioPlayerController
-import org.wycliffeassociates.otter.jvm.controls.controllers.ScrollSpeed
 import org.wycliffeassociates.otter.jvm.controls.waveform.WaveformImageBuilder
 import org.wycliffeassociates.otter.jvm.controls.model.VerseMarkerModel
 import org.wycliffeassociates.otter.jvm.device.audio.AudioConnectionFactory
@@ -45,11 +44,12 @@ import java.lang.Integer.min
 import org.wycliffeassociates.otter.jvm.controls.model.SECONDS_ON_SCREEN
 import kotlin.math.max
 import org.wycliffeassociates.otter.jvm.controls.model.ChunkMarkerModel
+import org.wycliffeassociates.otter.jvm.controls.waveform.IMarkerViewModel
 
 private const val WAV_COLOR = "#0A337390"
 private const val BACKGROUND_COLOR = "#FFFFFF"
 
-class VerseMarkerViewModel : ViewModel() {
+class VerseMarkerViewModel : ViewModel(), IMarkerViewModel {
 
     private val logger = LoggerFactory.getLogger(VerseMarkerViewModel::class.java)
 
@@ -66,16 +66,15 @@ class VerseMarkerViewModel : ViewModel() {
 
     lateinit var waveformMinimapImageListener: ChangeListener<Image>
 
-    var markerStateProperty = SimpleObjectProperty<VerseMarkerModel>()
-    val currentMarkerNumberProperty = SimpleIntegerProperty(0)
+    override val currentMarkerNumberProperty = SimpleIntegerProperty(0)
 
-    lateinit var markerModel: VerseMarkerModel
-    val markers = observableListOf<ChunkMarkerModel>()
-    val markerCountProperty = markers.sizeProperty
+    override var markerModel: VerseMarkerModel? = null
+    override val markers = observableListOf<ChunkMarkerModel>()
+    override val markerCountProperty = markers.sizeProperty
 
-    var audioController: AudioPlayerController? = null
+    override var audioController: AudioPlayerController? = null
 
-    val audioPlayer = SimpleObjectProperty<IAudioPlayer>()
+    override val audioPlayer = SimpleObjectProperty<IAudioPlayer>()
 
     val isLoadingProperty = SimpleBooleanProperty(false)
     val isPlayingProperty = SimpleBooleanProperty(false)
@@ -83,12 +82,12 @@ class VerseMarkerViewModel : ViewModel() {
     val headerTitle = SimpleStringProperty()
     val headerSubtitle = SimpleStringProperty()
     val compositeDisposable = CompositeDisposable()
-    val positionProperty = SimpleDoubleProperty(0.0)
-    var imageWidthProperty = SimpleDoubleProperty()
+    override val positionProperty = SimpleDoubleProperty(0.0)
+    override var imageWidthProperty = SimpleDoubleProperty()
 
     private var sampleRate: Int = 0 // beware of divided by 0
     private var totalFrames: Int = 0 // beware of divided by 0
-    private var resumeAfterScroll = false
+    override var resumeAfterScroll = false
 
     fun onDock() {
         isLoadingProperty.set(true)
@@ -120,7 +119,9 @@ class VerseMarkerViewModel : ViewModel() {
         markerCountProperty.onChangeAndDoNow {
             markerRatioProperty.set("$it/$totalMarkers")
         }
-        markers.setAll(markerModel.markers)
+        markerModel?.let { markerModel ->
+            markers.setAll(markerModel.markers)
+        }
     }
 
     private fun loadTitles() {
@@ -132,18 +133,7 @@ class VerseMarkerViewModel : ViewModel() {
     private fun writeMarkers(): Completable {
         audioPlayer.get()?.pause()
         audioPlayer.get()?.close()
-        return markerModel.writeMarkers()
-    }
-
-    fun calculatePosition() {
-        audioPlayer.get()?.let { audioPlayer ->
-            val current = audioPlayer.getLocationInFrames()
-            val duration = audioPlayer.getDurationInFrames().toDouble()
-            val percentPlayed = current / duration
-            val pos = percentPlayed * imageWidthProperty.value
-            positionProperty.set(pos)
-            updateCurrentPlaybackMarker(current)
-        }
+        return markerModel?.writeMarkers() ?: Completable.complete()
     }
 
     fun saveAndQuit() {
@@ -166,33 +156,6 @@ class VerseMarkerViewModel : ViewModel() {
         }
     }
 
-    fun placeMarker() {
-        markerModel.addMarker(audioPlayer.get().getLocationInFrames())
-        markers.setAll(markerModel.markers)
-    }
-
-    fun seekNext() {
-        val wasPlaying = audioPlayer.get().isPlaying()
-        if (wasPlaying) {
-            audioController?.toggle()
-        }
-        seek(markerModel.seekNext(audioPlayer.get().getLocationInFrames()))
-        if (wasPlaying) {
-            audioController?.toggle()
-        }
-    }
-
-    fun seekPrevious() {
-        val wasPlaying = audioPlayer.get().isPlaying()
-        if (wasPlaying) {
-            audioController?.toggle()
-        }
-        seek(markerModel.seekPrevious(audioPlayer.get().getLocationInFrames()))
-        if (wasPlaying) {
-            audioController?.toggle()
-        }
-    }
-
     fun initializeAudioController(slider: Slider) {
         audioController = AudioPlayerController(slider)
         audioController?.load(audioPlayer.get())
@@ -201,48 +164,6 @@ class VerseMarkerViewModel : ViewModel() {
 
     fun pause() {
         audioController?.pause()
-    }
-
-    private fun isPlaying(): Boolean {
-        return audioController?.isPlayingProperty?.value ?: false
-    }
-
-    fun rewind(speed: ScrollSpeed) {
-        if (isPlaying()) {
-            resumeAfterScroll = true
-            mediaToggle()
-        }
-        audioController?.rewind(speed)
-    }
-
-    fun fastForward(speed: ScrollSpeed) {
-        if (isPlaying()) {
-            resumeAfterScroll = true
-            mediaToggle()
-        }
-        audioController?.fastForward(speed)
-    }
-
-    fun resumeMedia() {
-        if (resumeAfterScroll) {
-            mediaToggle()
-            resumeAfterScroll = false
-        }
-    }
-
-    fun mediaToggle() {
-        if (audioController?.isPlayingProperty?.value == false) {
-            /* trigger change to auto-scroll when it starts playing */
-            val currentMarkerIndex = currentMarkerNumberProperty.value
-            currentMarkerNumberProperty.set(-1)
-            currentMarkerNumberProperty.set(currentMarkerIndex)
-        }
-        audioController?.toggle()
-    }
-
-    fun seek(location: Int) {
-        audioController?.seek(location)
-        updateCurrentPlaybackMarker(location)
     }
 
     private fun createWaveformImages(audio: AudioFile) {
@@ -289,14 +210,6 @@ class VerseMarkerViewModel : ViewModel() {
         return pixelsInDuration.toDouble()
     }
 
-    fun getLocationInFrames(): Int {
-        return audioPlayer.get().getLocationInFrames() ?: 0
-    }
-
-    fun getDurationInFrames(): Int {
-        return audioPlayer.get().getDurationInFrames() ?: 0
-    }
-
     fun pixelsInHighlight(controlWidth: Double): Double {
         if (sampleRate == 0 || totalFrames == 0) {
             return 1.0
@@ -305,26 +218,5 @@ class VerseMarkerViewModel : ViewModel() {
         val framesInHighlight = sampleRate * SECONDS_ON_SCREEN
         val framesPerPixel = totalFrames / max(controlWidth, 1.0)
         return max(framesInHighlight / framesPerPixel, 1.0)
-    }
-
-    private fun updateCurrentPlaybackMarker(currentFrame: Int) {
-        val currentMarkerFrame = markerModel.seekCurrent(currentFrame)
-        val currentMarker = markers.find { it.frame == currentMarkerFrame }
-        val index = currentMarker?.let { markers.indexOf(it) } ?: 0
-        currentMarkerNumberProperty.set(index)
-    }
-
-    fun requestAudioLocation(): Int {
-        return audioPlayer.value?.getLocationInFrames() ?: 0
-    }
-
-    fun undoMarker() {
-        markerModel.undo()
-        markers.setAll(markerModel.markers)
-    }
-
-    fun redoMarker() {
-        markerModel.redo()
-        markers.setAll(markerModel.markers)
     }
 }
